Change datatype from double to float

+++ src/classical/linear_regression/c_backend/linear_regression.h
@@ -1,8 +1,8 @@
 #ifndef LINEAR_REGRESSION_H
 #define LINEAR_REGRESSION_H
 
-void fit(double *X, double *Y, double *Beta, int num_samples, int num_input_features, int num_output_features);
-void predict(double *X, double *Beta, double *Prediction, int num_samples, int num_input_features, int num_output_features);
-double cost(double *Y_pred, double *Y, int num_samples, int num_output_features);
+void fit(float *X, float *Y, float *Beta, int num_samples, int num_input_features, int num_output_features);
+void predict(float *X, float *Beta, float *Prediction, int num_samples, int num_input_features, int num_output_features);
+float cost(float *Y_pred, float *Y, int num_samples, int num_output_features);
 
 #endif /* LINEAR_REGRESSION_H */
